+++ visual_behavior/validation/extended_trials.py
@@ -3,7 +3,7 @@ import pandas as pd
 
 from .trials import validate_trials
 from ..schemas.extended_trials import ExtendedTrialSchema
-from ..validation import assert_is_valid_dataframe
+from .utils import assert_is_valid_dataframe
 
 
 def validate_schema(extended_trials):
